refactored Multifaceted Abstraction code smell and extracted a few methods in to a separate class
